==> code/01a_imdb_scrape.ipynb <==
# Import libaries
import pandas as pd
import requests
from bs4 import BeautifulSoup

#List of genres from IMDB 

imdb_genres_list = [
    'Action',
    'Adventure',
    'Animation',
    'Biography',
    'Comedy',
    'Crime',
    'Documentary',
    'Drama',
    'Family',
    'Fantasy',
    'Game-Show',
    'History',
    'Horror',
    'Music',
    'Musical',
    'Mystery',
    'News',
    'Reality-TV',
    'Romance',
    'Sci-Fi',
    'Short',
    'Sport',
    'Talk-Show',
    'Thriller',
    'War',
    'Western'
]

titles=[]
entries_wanted = 500  # Pages load in increments of 50

for input_genre in imdb_genres_list:
    start_index = 1
    while start_index < entries_wanted:
        url = f'https://www.imdb.com/search/title/?title_type=tv_series,tv_miniseries&num_votes=5000,&genres={input_genre}&sort=num_votes,desc&start={start_index}&ref_=adv_nxt'
        res = requests.get(url)
        soup = BeautifulSoup(res.text)

        for list_item in soup.find_all('div', attrs={'class':'lister-item mode-advanced'}):
            link = list_item.find_all('a')[1]
            years = list_item.find('span', attrs={'class':'lister-item-year text-muted unbold'})
            description_obj = list_item.find_all('p', attrs={'class':'text-muted'})
            pg_rating = list_item.find('span', attrs={'class':'certificate'})
            genres = list_item.find('span', attrs={'class':'genre'})
            rating = list_item.find('meta', attrs={'itemprop':'ratingValue'})
            num_votes = list_item.find('meta', attrs={'itemprop':'ratingCount'})
            img_thumbnail = list_item.find('img')
            try:
                entry = {
                    'name': link.text,
                    'href': link['href'].strip(),
                    'years': years.text,
                    # 'verification': i.find('h3').text.strip().replace('\n', ' '),
                    # 'input_genre': input_genre,
                    'imdb_description': description_obj[1].text.strip().replace('\n', ' '),
                    'pg_rating': pg_rating.text.strip(),
                    'imdb_genre_tags': genres.text.strip(),
                    'imdb_rating': float(rating['content']),
                    'num_votes': int(num_votes['content']),
                    'img_thumbnail': img_thumbnail['loadlate']
                }
            except:
                pass
            
            titles.append(entry)
        
        start_index += 50


# Technique reference: https://stackoverflow.com/questions/51996139/grabbing-meta-content-with-beautiful-soup

len(titles)

titles_df = pd.DataFrame(titles)
titles_df.drop_duplicates(inplace=True)
titles_df

titles_df.to_csv('../data/imdb_titles_df.csv', index=False)

no_dupes = titles_df.drop_duplicates(subset='href')
no_dupes.shape

no_dupes.to_csv('../data/imdb_titles_no_dupes.csv', index=False)
