==> code/01b_wikipedia_scrape.ipynb <==
# Import libaries
import numpy as np
import pandas as pd
import requests
from tqdm import tqdm

### Read in table with titles to be scraped

no_dupes = pd.read_csv('../data/imdb_titles_no_dupes.csv')

no_dupes.shape

no_dupes.head(10)

### Define functions for our API call

def wiki_text_query(inp_subject, start_year=""):
    '''a wrappper for a function call to wiki_api_query_internal(subject) that first gets the "correct" search term'''

    warning_flag, search_string, search_url = wiki_search_term(inp_subject, start_year)

    try:
        output = wiki_text_query_internal(search_string)
        if len(output) == 0:
            print(f"Warning: 0 length output for: {search_string}")
        
        return warning_flag, search_string, search_url, output

    except:
        print(f"ERROR 2: broken search for: {search_string}")
        return warning_flag, "", "", ""
    

def wiki_search_term(inp_subject, start_year=""):
    '''returns the "correct" search term to use, that first tries to search for _____({year } TV_series) 
        then tries _____(TV_series), before trying ______
        Warning Flag codes:
        0 - page found under _____({year } TV_series)
        1 - page found under _____(TV_series)
        2 - page found under _____
        3 - no page found for search term
        4 - page found using manual override search term
        '''

    warningflag = 0
    
    S = requests.Session()
    URL = "https://en.wikipedia.org/w/api.php"
    
    PARAMS = {
        "action": "opensearch",
        "namespace": "0",
        "limit": "5",
        "format": "json"
    }

    R = S.get(url=URL, params=PARAMS | {'search': inp_subject + f" ({start_year} TV series)"})
    DATA = R.json()
    # print(f"DATA: {DATA}")

    if len(DATA[1]) > 0:
        return warningflag, DATA[1][0], DATA[3][0]
    else:
        R = S.get(url=URL, params=PARAMS | {'search': inp_subject + " (TV series)"})
        DATA = R.json()
        warningflag = 1

        if len(DATA[1]) > 0:
            return warningflag, DATA[1][0], DATA[3][0]
        else:
            R = S.get(url=URL, params=PARAMS | {'search': inp_subject})
            DATA = R.json()
            warningflag = 2

            if len(DATA[1]) > 0:
                return warningflag, DATA[1][0], DATA[3][0]
            else:
                warningflag = 3
                return warningflag, DATA[0], ""

# Reference: https://www.mediawiki.org/wiki/API:Opensearch

def wiki_text_query_internal(subject):

    url = 'https://en.wikipedia.org/w/api.php'

    params = {
            'action': 'query',
            'format': 'json',
            'titles': subject,
            'prop': 'extracts',
            # 'exintro': True,      #if True- returns just the first blurb on page
            'explaintext': True,
            'redirects':True,
        }
    
    response = requests.get(url, params=params)
    data = response.json()
    
    page = next(iter(data['query']['pages'].values()))

    return page['extract']

    
# Reference: https://www.jcchouinard.com/wikipedia-api/
# Reference: https://www.mediawiki.org/wiki/API:Parsing_wikitext
# Reference: https://www.mediawiki.org/wiki/API:Get_the_contents_of_a_page

s = requests.Session()
s.auth = ('user', 'pass')
s.headers.update({'x-test': 'true'})

# both 'x-test' and 'x-test2' are sent
s.get('http://httpbin.org/headers', headers={'x-test2': 'true'})

def override_query(inp_subject, start_year=""):
    '''direct search without trying alternate terms'''

    warning_flag = 4
    search_string = inp_subject
    
    search_url = f"https://en.wikipedia.org/wiki/{inp_subject.replace(' ', '_')}"

    try:
        output = wiki_text_query_internal(search_string)
        if len(output) == 0:
            print(f"Warning: 0 length output for: {search_string}")
        
        return warning_flag, search_string, search_url, output

    except:
        print(f"ERROR 2: broken search for: {search_string}")
        return warning_flag, "", "", ""

# Reference: https://www.jcchouinard.com/wikipedia-api/
# Reference: https://www.mediawiki.org/wiki/API:Parsing_wikitext
# Reference: https://www.mediawiki.org/wiki/API:Get_the_contents_of_a_page

### API query for the article text of each entry in our list

# Create an array of the start year of each series, to help handle edge cases where the year is contained in the URL

start_years = {}
for i in range(len(no_dupes)):
    year_entry = no_dupes['years'][i]
    temp_read_in = year_entry[1:5]
    if (temp_read_in[0] != '1') and (temp_read_in[0] != '2'):
        j = 0
        while j < len(year_entry):
            if (year_entry[j] == '1') or (year_entry[j] == '2'):
                break
            j += 1
        start_index = j-1
        temp_read_in = year_entry[j: j+4]
    
    start_years[i] = temp_read_in

wiki_articles = {}

for i in tqdm(range(len(no_dupes))):

    name = no_dupes['name'][i]
    start_year = start_years[i]
    
    warning_flag, search_term, search_url, wiki_text = wiki_text_query(name, start_year)
    
    entry = {
        'index_no': i,
        'name' : name,
        'wiki_warning_flag' : warning_flag,
        'wiki_search_term' : search_term,
        'wiki_search_url': search_url,
        'wiki_text' : wiki_text
    }

    wiki_articles[i] = entry

wiki_articles[0]

### Create a DataFrame from the extracted data

wiki_articles_df = pd.DataFrame.from_dict(wiki_articles, orient='index')

wiki_articles_df.head()

#double-checking no entry count mismatch
len(wiki_articles_df)

### Identifying entries that did not run as intended

wiki_articles_df['wiki_warning_flag'].value_counts()

##### Blank entries:

wiki_articles_df[wiki_articles_df['wiki_text'] == '']

Empty Entries:
> Tokyo Ghoul: Root A ;	
Rurouni Kenshin: Trust and Betrayal ;	
The Wheel of Time: Origins ;	
Owari no serafu	;
Scam 1992: The Harshad Mehta Story ;	
Self Made: Inspired by the Life of Madam C.J. Walker ;	
The Filthy Frank Show ;	
Inside Look: The Assassination of Gianni Versace - American Crime Story ;	
Jeen-yuhs: A Kanye Trilogy ;	
FIFA Uncovered ;	
I Just Killed My Dad ;	
Sins of Our Mother ;	
A Young Doctor's Notebook & Other Stories ;	
Pera Palas'ta Gece Yarisi ;	
The Heavy Water War: Stopping Hitler's Atomic Bomb ;	
Doctor Foster: A Woman Scorned ;

# Create a dictionary mapping from old search term to new (override) search term

search_term_transform = {
    "Tokyo Ghoul: Root A" : "Tokyo Ghoul",
    "Rurouni Kenshin: Trust and Betrayal" : "Rurouni Kenshin: Trust & Betrayal",
    "The Wheel of Time: Origins" : "The Wheel of Time",
    "Owari no serafu" : "Seraph of the End",
    "Scam 1992: The Harshad Mehta Story" : "Scam 1992",
    "Self Made: Inspired by the Life of Madam C.J. Walker" : "Self Made (miniseries)",
    "The Filthy Frank Show" : "Joji (musician)",
    "Inside Look: The Assassination of Gianni Versace - American Crime Story" : "The Assassination of Gianni Versace: American Crime Story",
    "Jeen-yuhs: A Kanye Trilogy" : "Jeen-Yuhs",
    # "FIFA Uncovered" : "",        # No applicable Wikipedia entry found
    # "I Just Killed My Dad" : "",  # No applicable Wikipedia entry found
    "Sins of Our Mother" : "Killings of Tylee Ryan and J. J. Vallow",
    "A Young Doctor's Notebook & Other Stories" : "A Young Doctor's Notebook (TV series)",
    "Pera Palas'ta Gece Yarisi" : "Midnight at the Pera Palace",
    "The Heavy Water War: Stopping Hitler's Atomic Bomb" : "The Heavy Water War",
    "Doctor Foster: A Woman Scorned" : "Doctor Foster (TV series)"
}

# Fixing the entries that came up empty

for name in search_term_transform:

    index_no = int(wiki_articles_df[wiki_articles_df['name'] == name]['index_no'])
    start_year = ""
    
    warning_flag, search_term, search_url, wiki_text = override_query(search_term_transform[name] )
    
    wiki_articles[index_no]['wiki_warning_flag'] = warning_flag
    wiki_articles[index_no]['wiki_search_term'] = search_term
    wiki_articles[index_no]['wiki_search_url'] = search_url
    wiki_articles[index_no]['wiki_text'] = wiki_text


#double-check for missing entries (2 expected)
wiki_articles_df = pd.DataFrame.from_dict(wiki_articles, orient='index')
len(wiki_articles_df[wiki_articles_df['wiki_text'] == ''])

wiki_articles_df[wiki_articles_df['wiki_text'] == '']

##### Entries containing the phrase "may refer to" means the wrong page (disambiguation) was scraped

may_refer_to_df = wiki_articles_df[wiki_articles_df['wiki_text'].str.contains("may refer to")]

may_refer_to_df.shape

may_refer_to_df.index

may_refer_to_df[['name']]

# Create a dictionary mapping from old search term to new (override) search term

search_term_transform2 = {
    'The Family Man' : 'The Family Man (Indian TV series)', 
    'The River': 'The River (U.S. TV series)', 
    'In the Dark': 'In the Dark (American TV series)', 
    'Law of the Lawless': 'Brigada', 
    'Tomorrow': 'Tomorrow (South Korean TV series)', 
    'Wrecked': 'Wrecked (American TV series)', 
    'Into the West': 'Into the West (miniseries)',
    'Big Mouth': 'Big Mouth (American TV series)', 
    'Food Wars': 'Food Wars!: Shokugeki no Soma', 
    'Gantz': 'Gantz', 
    'Fighting Spirit': 'Hajime no Ippo', 
    'The Liberator': 'The Liberator (miniseries)',
    'Golden Boy': 'Golden Boy (manga)', 
    'When They Cry': 'Higurashi When They Cry', 
    '86': '86 (novel_series)', 
    'Carlos': 'Carlos (miniseries)', 
    'Aquarius': 'Aquarius (American TV series)', 
    'Louie': 'Louie (American TV series)', 
    'Peep Show': 'Peep Show (British TV series)', 
    'Wilfred': 'Wilfred (American TV series)', 
    'Last Man Standing': 'Last Man Standing (American TV series)', 
    'Rescue Me': 'Rescue Me (American TV series)',
    # 'Being Human': 'Being Human (North American TV series)',  #gives error- handle separately
    'Crashing': 'Crashing (British TV series)', 
    'The Kingdom': 'The Kingdom (miniseries)', 
    'The Killing': 'The Killing (U.S. TV series)', 
    'The Outsider': 'The Outsider (miniseries)',
    'The Stranger': 'The Stranger (British TV series)', 
    'Leverage': 'Leverage (American TV series)', 
    'Big Sky': 'Big Sky (American TV series)', 
    'True Story': 'True Story (miniseries)', 
    # 'Shameless': 'Shameless (British TV series)',  #gives error- handle separately
    'The Finder': 'The Finder (American TV series)',
    'Wallander': 'Wallander (British TV series)',
    'Black Sun': 'Balkan Shadows',
    'Bordertown': 'Bordertown (Finnish TV series)',
    'Ashes to Ashes': 'Ashes to Ashes (British TV series)',
    'Band of Brothers': 'Band of Brothers (miniseries)',
    'Heroes': 'Heroes (American TV series)',
    'Dopesick': 'Dopesick (miniseries)',
    'The Magicians': 'The Magicians (American TV series)',
    'Little Fires Everywhere': 'Little Fires Everywhere (miniseries)',
    'Messiah': 'Messiah (American TV series)',
    'Heartland': 'Heartland (Canadian TV series)',
    'Shake It Up': 'Shake It Up (American TV series)',
    'Wipeout': 'Wipeout (2008 game show)',
    # 'Rebelde': 'Rebelde',  #gives error- handle separately
    'The Gift': 'The Gift (Turkish TV series)',
    # 'The Returned': '',  # index 1956 /title/tt2521668/	The Returned (French TV series)
    'The Shining': 'The Shining (TV miniseries)',
    # 'The Returned': '', # index 1981 /title/tt3230780/	The Returned (American TV series)
    'The Circle': 'The Circle (American TV series)',
    'Face Off': 'Face Off (American TV series)',
    'Rose Red': 'Rose Red (miniseries)',
    # 'Taken': 'Taken (miniseries)',  #gives error- handle separately
    'Echoes': 'Echoes (miniseries)',
    'Flower of Evil': 'Flower of Evil (South Korean TV series)',
    'North & South': 'North & South (TV serial)',
    'The L Word': 'The L Word',
    'The Thorn Birds': 'The Thorn Birds (miniseries)',
    'True Beauty': 'True Beauty (South Korean TV series)',
    'Fingersmith': 'Fingersmith (TV serial)',
    'Still Standing': 'Still Standing (American TV series)',
    'Run': 'Run (American TV series)',
    'The View': 'The View (talk show)',
    'What/If': 'What/If'
}

# Fixing the entries that grabbed data from the disambiguation page

for name in search_term_transform2:
    # print(f"{name}; {search_term_transform2[name]}")  #used to identify outlier cases that need to be fixed separately

    index_no = int(wiki_articles_df[wiki_articles_df['name'] == name]['index_no'])
    start_year = ""
    
    warning_flag, search_term, search_url, wiki_text = override_query(search_term_transform2[name] )
    
    wiki_articles[index_no]['wiki_warning_flag'] = warning_flag
    wiki_articles[index_no]['wiki_search_term'] = search_term
    wiki_articles[index_no]['wiki_search_url'] = search_url
    wiki_articles[index_no]['wiki_text'] = wiki_text

#Outlier cases
bad_indices = [1218, 1506, 1931, 1956, 1981, 2193]
search_term_transform3 = ["Being Human (North American TV series)",
                          "Shameless (British TV series)",
                          "Rebelde",
                          "The Returned (French TV series)", 
                          "The Returned (American TV series)",
                          "Taken (miniseries)"]

for i in range(len(bad_indices)):
    
    index_no = bad_indices[i]
    start_year = ""
    
    warning_flag, search_term, search_url, wiki_text = override_query(search_term_transform3[i] )
    
    wiki_articles[index_no]['wiki_warning_flag'] = warning_flag
    wiki_articles[index_no]['wiki_search_term'] = search_term
    wiki_articles[index_no]['wiki_search_url'] = search_url
    wiki_articles[index_no]['wiki_text'] = wiki_text

#double-check for entries still pointing to disambiguation page
wiki_articles_df = pd.DataFrame.from_dict(wiki_articles, orient='index')
len(wiki_articles_df[wiki_articles_df['wiki_text'].str.contains("may refer to")])

#### Summarize & write output to csv/json

wiki_articles_df.head()

wiki_articles_df.shape

wiki_articles_df['wiki_warning_flag'].value_counts()

wiki_articles_df.to_csv('../data/wiki_articles_df.csv', index=False)

wiki_articles_df.to_json('../data/wiki_articles_df.json')  #backup, in case the commas inside 'wiki_text' column throw off comma separation in csv

# test_read_in_csv = pd.read_csv('../data/wiki_articles_df.csv', index_col="index_no")
# test_read_in_csv.head()


# test_read_in_json = pd.read_json('../data/wiki_articles_df.json')
# test_read_in_json.head()
