==> code/01c_themoviedb_scrape.ipynb <==
# Import libaries
import numpy as np
import pandas as pd
import requests
from bs4 import BeautifulSoup
from urllib.request import Request, urlopen
from tqdm import tqdm

# Read in API key for TMDB

with open('../ignored/tmdb_api_key.txt', 'r') as file:
    my_api_key = file.read().rstrip()

# Technique reference: https://stackoverflow.com/questions/8369219/how-to-read-a-text-file-into-a-string-variable-and-strip-newlines

### Read in table with titles to be scraped

data_directory = "../data/"

no_dupes = pd.read_csv(f'{data_directory}imdb_titles_no_dupes.csv')
no_dupes.shape

no_dupes.head()

### Define functions for our API call

def tmdb_find(external_id, api_key=my_api_key):

    base_url = 'https://api.themoviedb.org/3/find/'
    url = f"{base_url}{external_id}?api_key={my_api_key}"

    params = {
            'language': 'en-US',
            'external_source': 'imdb_id'
        }
    
    response = requests.get(url, params=params)
    data = response.json()['tv_results'][0]

    data_to_keep = {
        'tmdb_id' : int(data['id']),
        'tmdb_name' : data['name'],
        'original_name' : data['original_name'],
        'original_language' : data['original_language'],        
        'origin_country' : data['origin_country'],
        'tmdb_popularity' : data['popularity'],
        'tmdb_vote_average' : data['vote_average'],
        'tmdb_vote_count' : int(data['vote_count']),
        'first_air_date' : data['first_air_date'],
        'tmdb_adult_content' : data['adult'],
        'tmdb_poster_path' : data['poster_path'],
        'tmdb_overview' : data['overview'].encode("ascii", "ignore").decode(), 
                            # Technique reference: https://pythonguides.com/remove-unicode-characters-in-python/
    }

    return data_to_keep

# Reference: https://developers.themoviedb.org/3/find/find-by-id

# query for other info not obtained from "find" search for tv show

def tmdb_tv_query(tv_id, api_key=my_api_key):

    base_url = 'https://api.themoviedb.org/3/tv/'
    url = f"{base_url}{tv_id}?api_key={my_api_key}"

    params = {
            'language': 'en-US',
        }
    
    response = requests.get(url, params=params)
    data = response.json()

    data_to_keep = {
        'tmdb_tagline' : data['tagline'],
        'tmdb_genres' : ", ".join( [i['name'] for i in data['genres']] ),
        'tv_networks' : ", ".join( [i['name'] for i in data['networks']] ),
    }

    return data_to_keep

# Reference: https://developers.themoviedb.org/3/tv/get-tv-details

def scrape_keywords(tmdb_id):
    
    base_url = 'https://www.themoviedb.org/tv/'
    full_url = base_url + str(tmdb_id)

    req = Request(full_url, headers={'User-Agent': 'Mozilla/5.0'})
    webpage = urlopen(req).read()
    soup = BeautifulSoup(webpage, "html.parser")

    # Technique reference to avoid 403 error: 
    # https://medium.com/@raiyanquaium/how-to-web-scrape-using-beautiful-soup-in-python-without-running-into-http-error-403-554875e5abed

    keywords_list = []
    keyword_sidebar = soup.find('section', attrs='keywords right_column')
    for li in keyword_sidebar.find_all('li'):
        keywords_list.append(li.text)

    return ", ".join(keywords_list)

def make_tmdb_entry(row_index, input_df):
    imdb_href = input_df['href'][row_index]
    entry = {'imdb_href' : imdb_href}
    external_id = imdb_href[7:-1]

    try:
        api_pull_dict = tmdb_find(external_id, my_api_key)
        tmdb_id = api_pull_dict['tmdb_id']

        entry.update(api_pull_dict)
        entry.update( tmdb_tv_query(tmdb_id, my_api_key) )
        entry['tmdb_keywords'] = scrape_keywords(tmdb_id)
    except:
        print(f"Error on row_index: {row_index} ; imdb_href: {imdb_href}")
        pass

    return entry

#### Create Dataframe of TMDB entries, based on list of shows from IMDB

tmdb_articles = {}

# for i in tqdm(range(len(no_dupes))):
for i in tqdm(range(len(no_dupes))):    
    tmdb_articles[i] = make_tmdb_entry(i, no_dupes)

tmdb_articles_df = pd.DataFrame.from_dict(tmdb_articles, orient='index')

tmdb_articles_df.tail()

tmdb_articles_df.shape

#### Handle outlier cases

tmdb_articles_df.isna().sum()

tmdb_articles_df[tmdb_articles_df['tmdb_genres'].isna()]['imdb_href']

#manual lookup for fixes

outlier_cases = {
    # row    imdb_href        tmdb_id
    # 48	:	""   , #   /title/tt1758429/
    50	 :  12971  , #       /title/tt0214341/
    121	 :  12609  , #       /title/tt0280249/
    # 161	:   ""     , #   /title/tt1598754/
    # 224	:  1010901 , #   /title/tt13623148/
    288	 : 36406   , #   /title/tt0249327/ 
    # 334	:  ""       , #  /title/tt14986406/
    # 368	  : 1085199 , #      /title/tt0387736/
    # 374	:   ""      , #  /title/tt5497534/
    420	 :  3570   , #   /title/tt0114327/
    # 467	:   ""     , #   /title/tt8213522/
    729	 :  114772 , #   /title/tt0063939/
    # 750	:   ""     , #   /title/tt0060019/
    # 765	:   ""     , #   /title/tt0062550/
    # 793	:   ""     , #   /title/tt0184175/
    819	:   31654  , #   /title/tt0211145/
    873	 :  81026  , #   /title/tt6601082/
    884	 :  29200  , #   /title/tt0086719/
    1041	:  135670 , #    /title/tt14650074/
    1103	:   7246  , #    /title/tt0096548/
    1115	:   7246  , #    /title/tt0088484/
    # 1182	:   ""    , #    /title/tt4202274/
    1384	:   1920  , #    /title/tt4093826/
    # 1553	:   ""     , #   /title/tt6111552/
    # 1554	:   ""    , #    /title/tt0090509/
    1624	:   64513  , #     /title/tt6205862/
    1640	:   64513  , #   /title/tt8010592/
    1718	:  208720  , #   /title/tt21867596/
    1924	:   13793  , #   /title/tt0094500/
    1940	:   42135  , #   /title/tt0389680/
    # 1983	:   ""     , #   /title/tt7686464/
    # 1995	:   ""     , #   /title/tt1635327/
    2066	:  131488  , #   /title/tt13720112/
    2178	:  134949  , #   /title/tt14153686/
    # 2216	:   ""     , #   /title/tt9566030/
}

len(outlier_cases)

def scrape_outlier(tmdb_id, my_api_key=my_api_key):

    # print(f"Attempting to scrape tmdb_id: {tmdb_id}")

    base_url = 'https://api.themoviedb.org/3/tv/'
    url = f"{base_url}{tmdb_id}?api_key={my_api_key}"

    params = {
            'language': 'en-US',
        }

    response = requests.get(url, params=params)
    data = response.json()

    entry = {}

    try:
        data_to_keep = {
            'tmdb_id' : int(data['id']),
            'tmdb_name' : data['name'],
            'original_name' : data['original_name'],
            'original_language' : data['original_language'],        
            'origin_country' : data['origin_country'],
            'tmdb_popularity' : data['popularity'],
            'tmdb_vote_average' : data['vote_average'],
            'tmdb_vote_count' : int(data['vote_count']),
            'first_air_date' : data['first_air_date'],
            'tmdb_adult_content' : data['adult'],
            'tmdb_poster_path' : data['poster_path'],
            'tmdb_overview' : data['overview'].encode("ascii", "ignore").decode(), 
        }
        
        entry.update(data_to_keep)
        entry.update( tmdb_tv_query(tmdb_id, my_api_key) )
        entry['tmdb_keywords'] = scrape_keywords(tmdb_id)
    except:
        
        data_to_keep = {
            'tmdb_id' : "",
            'tmdb_name' : "",
            'original_name' : "",
            'original_language' : "",      
            'origin_country' : "",
            'tmdb_popularity' : "",
            'tmdb_vote_average' : "",
            'tmdb_vote_count' : "",
            'first_air_date' : "",
            'tmdb_adult_content' : "",
            'tmdb_poster_path' : "",
            'tmdb_overview' : "",
            'tmdb_tagline' : "",
            'tmdb_genres' : "",
            'tv_networks' : "",
            'tmdb_keywords' : "",
        }
        entry.update(data_to_keep)

    return entry

for i in tqdm(outlier_cases):
    tmdb_articles[i].update( scrape_outlier(outlier_cases[i], my_api_key) )

tmdb_articles_df = pd.DataFrame.from_dict(tmdb_articles, orient='index')

tmdb_articles_df.fillna("", inplace=True)

#### Write finished dataframe to disk

tmdb_articles_df.to_csv('../data/tmdb_articles_df.csv', index=False)

# image_base_url = 'https://image.tmdb.org/t/p/original'
# test_image_url = image_base_url + asdf['poster_path']
# test_image_url
