==> code/02_merge_dataframes.ipynb <==
import pandas as pd

data_directory = "../data/"

#### Read in IMDB data

no_dupes = pd.read_csv(f'{data_directory}imdb_titles_no_dupes.csv')
no_dupes.head()

no_dupes.columns

no_dupes.shape

#### Read in Wikipedia data

wiki_articles_df = pd.read_csv(f'{data_directory}wiki_articles_df.csv', index_col="index_no")
wiki_articles_df.head()

wiki_articles_df.columns

wiki_articles_df.drop(columns=['name', 'wiki_warning_flag'], inplace=True)

wiki_articles_df.shape

#### Read in TheMovieDB data

tmdb_articles_df = pd.read_csv(f'{data_directory}tmdb_articles_df.csv')
tmdb_articles_df.head()

tmdb_articles_df.columns

tmdb_articles_df.drop(columns='imdb_href', inplace=True)

tmdb_articles_df.shape

#### Merge dataframes

merged_df = pd.concat([no_dupes, wiki_articles_df, tmdb_articles_df], axis=1)
merged_df.shape

merged_df.head()

merged_df.columns

merged_df.fillna("", inplace=True)

#### Creating a 'is_anime' column

# merged_df[merged_df['imdb_genre_tags'].str.contains("Animation")] # [c] 481

# merged_df[merged_df['tmdb_keywords'].str.contains("anime")] # [a] 191
# merged_df[merged_df['tmdb_keywords'].str.contains("anime") & merged_df['imdb_genre_tags'].str.contains("Animation")] # [a & c] 190
# merged_df[merged_df['tmdb_keywords'].str.contains("anime") & ~merged_df['imdb_genre_tags'].str.contains("Animation")] #[a & ~c] 1 - live-action Cowboy Bebop

# merged_df[merged_df['tmdb_keywords'].str.contains("anime") & merged_df['origin_country'].str.contains("JP")] # [a & b] 180
# merged_df[merged_df['tmdb_keywords'].str.contains("anime") & ~merged_df['origin_country'].str.contains("JP")] #[a & ~b] 11 - Western cartoons in anime style

# merged_df[merged_df['origin_country'].str.contains("JP")] # [b] 184
# merged_df[merged_df['origin_country'].str.contains("JP") & merged_df['imdb_genre_tags'].str.contains("Animation")] # [b & c] 182
# merged_df[merged_df['origin_country'].str.contains("JP")
#          & merged_df['imdb_genre_tags'].str.contains("Animation")
#          & ~merged_df['tmdb_keywords'].str.contains("anime")
#         ] # 2 - Yu-Gi-Oh & Battle Angel - [~a & b & c] both should be anime

# merged_df['wiki_text'].str.contains("anime")  # [d] 261
# merged_df['tmdb_genres'].str.contains("Animation") # [e] 465

a = merged_df['tmdb_keywords'].str.contains("anime") #191
b = merged_df['origin_country'].str.contains("JP") #184
c = merged_df['imdb_genre_tags'].str.contains("Animation") #481
d = merged_df['wiki_text'].str.contains("anime")  #261
# e = merged_df['tmdb_genres'].str.contains("Animation") #465

# merged_df[a & c & ~b]

merged_df['is_anime'] = 0
merged_df.loc[a&c, 'is_anime'] = 1

#### Write to disk

merged_df.to_csv(f'{data_directory}merged_df.csv', index=False)

merged_df.to_json(f'{data_directory}merged_df.json')

# test_read_in_csv = pd.read_csv(f'{data_directory}merged_df.csv')
# test_read_in_csv.head()
