==> code/03_text_preprocessing_unweighted_nowiki.ipynb <==
# Data Prep & Processing - Lemmatizing & Tokenizing

#### Set parameters for how output will be saved

data_directory = "../data/"

model_name_suffix = "unweighted_nowiki"
run_number = 1

tokenized_run_name = f"{model_name_suffix}_{run_number}_tokens"

#### Imports

# general use
import numpy as np
import pandas as pd
# import matplotlib.pyplot as plt

# data preparation for NLP / modeling
import nltk
from nltk.tokenize import RegexpTokenizer
from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords, wordnet
# from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

# # for LDA modeling & presentation
# from sklearn.model_selection import GridSearchCV
# from sklearn.decomposition import LatentDirichletAllocation
# import pyLDAvis
# import pyLDAvis.sklearn
# pyLDAvis.enable_notebook()

# # for LSA modeling
# from sklearn.decomposition import TruncatedSVD

# for progress bar on loops
from tqdm import tqdm

#### Data Read In

merged_df = pd.read_json(f'{data_directory}merged_df.json')
merged_df.head()

merged_df.isna().sum()

merged_df.fillna("",inplace=True)

#### Create separate DataFrame with only select columns to be used in lemmatizing/tokenizing

# custom weighting factors

pg_weighting_factor = 1            # =1 is unweighted
imdb_genre_weighting_factor = 1    # =1 is unweighted
tmdb_overview_weighting_factor = 1    # =1 is unweighted
tmdb_tagline_weighting_factor = 1    # =1 is unweighted
tmdb_keywords_weighting_factor = 1    # =1 is unweighted

# create copied DataFrame with only name and a combined column of text
df = merged_df[['name', 'href']].copy()
df['combined_text'] = (merged_df['imdb_description'] + " . " \
                        + (merged_df['pg_rating'].replace("-","") + ", ") * pg_weighting_factor + " . " \
                        + (merged_df['imdb_genre_tags'] + ", ") * imdb_genre_weighting_factor + " . " \
                        + (merged_df['tmdb_overview'] + ", ") * tmdb_overview_weighting_factor + " . " \
                        + (merged_df['tmdb_tagline'] + ", ") * tmdb_tagline_weighting_factor + " . " \
                        + (merged_df['tmdb_keywords'] + ", ") * tmdb_keywords_weighting_factor + " . " \
                        # + merged_df['wiki_text'] + " . "
                    ).replace('==', '').replace('\n', '.').replace("\\'s", "'s").replace("\'s", "'s")  #seems like some of these don't work?
df.head()

## Data prep - Text normalization (lemmatizing, tokenizing, stop words)

#### Lemmatizing

# Lemmatizing Helper Function (lemmatize individual words)
# Reference: General Assembly DSI Lesson 504-lesson-nlp-i

def custom_lemmatize(word, tag):
    
    wn = WordNetLemmatizer()
    
    mapper = {
        'J': wordnet.ADJ,
        'V': wordnet.VERB,
        'N': wordnet.NOUN,
        'R': wordnet.ADV
    }
    pos = mapper.get(tag[0])
    
    return wn.lemmatize(word, pos) if pos else word

# Lemmatizing 
# Reference: General Assembly DSI Lesson 504-lesson-nlp-i

def lemmatize_func(df_to_process, input_text_col='combined_text', output_col_name='lemmatized_text'):

    df_to_process[output_col_name] = ""

    for i in tqdm(range(df_to_process.shape[0])):
        df_to_process.loc[i, output_col_name] = \
            " ".join([custom_lemmatize(word, tag) for word, tag in nltk.pos_tag( df_to_process[input_text_col][i].split(" "))])
    
    return

lemmatize_func(df, 'combined_text', 'lemmatized_text')

#### Tokenizing & stop word removal

# Tokenizing & stopword removal AFTER lemmatizing
# Reference: General Assembly DSI Lesson 504-lesson-nlp-i

def tokenize_func(df_to_process, input_text_col='lemmatized_text', output_col_name='tokenized_text'):

    tokenizer = RegexpTokenizer('\w+')

    df_to_process[output_col_name] = ""

    for i in tqdm(range(df.shape[0])):
        my_text_tokens = tokenizer.tokenize(df_to_process[input_text_col][i].lower())
        my_text_tokens_ns = [token for token in my_text_tokens if token not in stopwords.words("english")]
        recombined_text = " ".join(my_text_tokens_ns)
        df_to_process.loc[i, output_col_name] = recombined_text
    
    return


tokenize_func(df, 'lemmatized_text', 'tokenized_text')

#### Preview & save Dataframe with only tokenized_text

df.head(3)

columns_to_save = ['name', 'href', 'tokenized_text']

df[columns_to_save].to_csv(f"{data_directory}{tokenized_run_name}.csv", index=False)
