==> code/03b_corpora_dictionary_02.ipynb <==
# Data Prep - Generating corpora & dictionary

#### Set parameters for how output will be saved

data_directory = "../data/"

input_filename = "unweighted_1_tokens"
model_name_suffix = ""

output_name = f"{model_name_suffix}corpora_dictionary_output"

#### Imports

# general use
import numpy as np
import pandas as pd
# import matplotlib.pyplot as plt

import gensim
import os
from gensim.utils import simple_preprocess
from gensim.test.utils import get_tmpfile
from gensim import corpora


from tqdm import tqdm

#### Data Read In

tokens_df = pd.read_csv(f'{data_directory}{input_filename}.csv')
tokens_df.head()

tokens_df.isna().sum()

def make_corpora(input_df, col_to_use):
    my_corpora =""

    for i in range(len(input_df)):
        try:
            rowtemp = input_df.loc[i, col_to_use].replace("\'s", "'s").replace('==', '').replace('\n', '.')#.replace("\\'s", "'s")

            my_corpora += rowtemp.encode("ascii", "ignore").decode()
                            # Technique reference: https://pythonguides.com/remove-unicode-characters-in-python/
        except:
            pass
            
    return my_corpora


# Technique reference: https://www.geeksforgeeks.org/nlp-gensim-tutorial-complete-guide-for-beginners/

my_corpora = make_corpora(tokens_df, "tokenized_text")
len(my_corpora)

# preprocess the file to get a list of tokens
tokenized =[]
for sentence in my_corpora.split('.'):
  # the simple_preprocess function returns a list of each sentence
  tokenized.append(simple_preprocess(sentence, deacc = True, min_len=1))
 
# print(tokenized)

# Technique reference: https://www.geeksforgeeks.org/nlp-gensim-tutorial-complete-guide-for-beginners/

my_dictionary = corpora.Dictionary(tokenized)
len(my_dictionary)

# Code/Technique excerpted directly from: https://www.geeksforgeeks.org/nlp-gensim-tutorial-complete-guide-for-beginners/

# save your dictionary to disk
# my_dictionary.save('my_dictionary.dict')
my_dictionary.save(f'{tmp_fname}.dict')
 
# load back
# load_dict = corpora.Dictionary.load('my_dictionary.dict')
                                   
# save your dictionary as text file

# tmp_fname = get_tmpfile("dictionary")
tmp_fname = output_name
my_dictionary.save_as_text(tmp_fname)
 
# load your dictionary text file
# load_dict = corpora.Dictionary.load_from_text(tmp_fname)
