==> code/03d_exploded_matrix.ipynb <==
import numpy as np
import pandas as pd
# import matplotlib.pyplot as plt

# from scipy import sparse
# from sklearn.metrics.pairwise import pairwise_distances, cosine_distances

# for progress bar on loops
from tqdm import tqdm

lookup_df = pd.read_csv(f"{raw_data_directory}merged_nowiki_df.csv")

df2[['team1','team2']] = pd.DataFrame(df2.teams.tolist(), index= df2.index)
print (df2)
